+++ python/ray/data/_internal/data_iterator/iterator_impl.py
@@ -3,7 +3,7 @@ from typing import TYPE_CHECKING, Optional, Union, Iterator, Tuple
 from ray.types import ObjectRef
 from ray.data.block import Block, BlockMetadata
 from ray.data.context import DataContext
-from ray.data.dataset_iterator import DataIterator
+from ray.data.datastream_iterator import DataIterator
 from ray.data._internal.stats import DatastreamStats
 
 if TYPE_CHECKING:
