+++ python/ray/data/_internal/data_iterator/stream_split_iterator.py
@@ -14,7 +14,7 @@ from typing import (
 
 import ray
 
-from ray.data.dataset_iterator import DataIterator
+from ray.data.datastream_iterator import DataIterator
 from ray.data.block import Block, BlockMetadata
 from ray.data.context import DataContext
 from ray.data._internal.execution.streaming_executor import StreamingExecutor
