+++ python/ray/data/_internal/pipeline_executor.py
@@ -6,14 +6,14 @@ import logging
 import ray
 from ray.data.block import T
 from ray.data.context import DataContext
-from ray.data.dataset import Datastream
+from ray.data.datastream import Datastream
 from ray.data._internal.progress_bar import ProgressBar
 from ray.data._internal import progress_bar
 
 logger = logging.getLogger(__name__)
 
 if TYPE_CHECKING:
-    from ray.data.dataset_pipeline import DatasetPipeline
+    from ray.data.datastream_pipeline import DatasetPipeline
 
 
 def pipeline_stage(fn: Callable[[], Datastream[T]]) -> Datastream[T]:
