+++ python/ray/data/dataset_pipeline.py
@@ -35,8 +35,8 @@ from ray.data._internal.plan import ExecutionPlan
 from ray.data._internal.stats import DatasetPipelineStats, DatastreamStats
 from ray.data.block import BatchUDF, Block, DataBatch, KeyFn, RowUDF, T, U
 from ray.data.context import DataContext
-from ray.data.dataset import Datastream
-from ray.data.dataset_iterator import DataIterator
+from ray.data.datastream import Datastream
+from ray.data.datastream_iterator import DataIterator
 from ray.data.datasource import Datasource
 from ray.data.datasource.file_based_datasource import (
     BlockWritePathProvider,
