+++ python/ray/data/tests/preprocessors/test_preprocessors.py
@@ -232,9 +232,9 @@ def test_transform_all_formats(create_dummy_preprocessors, pipeline, dataset_for
         _apply_transform(with_nothing, ds)
 
     if pipeline:
-        patcher = patch.object(ray.data.dataset_pipeline.DatasetPipeline, "map_batches")
+        patcher = patch.object(ray.data.datastream_pipeline.DatasetPipeline, "map_batches")
     else:
-        patcher = patch.object(ray.data.dataset.Datastream, "map_batches")
+        patcher = patch.object(ray.data.datastream.Datastream, "map_batches")
 
     with patcher as mock_map_batches:
         _apply_transform(with_pandas, ds)
