+++ release/nightly_tests/dataset/iter_batches_benchmark.py
@@ -2,7 +2,7 @@ import sys
 from typing import Optional
 
 import ray
-from ray.data.dataset import Dataset
+from ray.data.datastream import Dataset
 
 from benchmark import Benchmark
