+++ release/nightly_tests/dataset/read_parquet_benchmark.py
@@ -1,5 +1,5 @@
 import ray
-from ray.data.dataset import Dataset
+from ray.data.datastream import Dataset
 
 from benchmark import Benchmark
 from parquet_data_generator import generate_data
